feat: merge staticcheck, stylecheck, gosimple
