--- pkg/golinters/gosimple/gosimple.go
@@ -1,22 +1,0 @@
-package gosimple
-
-import (
-	"honnef.co/go/tools/simple"
-
-	"github.com/golangci/golangci-lint/pkg/config"
-	"github.com/golangci/golangci-lint/pkg/goanalysis"
-	"github.com/golangci/golangci-lint/pkg/golinters/internal"
-)
-
-func New(settings *config.StaticCheckSettings) *goanalysis.Linter {
-	cfg := internal.StaticCheckConfig(settings)
-
-	analyzers := internal.SetupStaticCheckAnalyzers(simple.Analyzers, cfg.Checks)
-
-	return goanalysis.NewLinter(
-		"gosimple",
-		"Linter for Go source code that specializes in simplifying code",
-		analyzers,
-		nil,
-	).WithLoadMode(goanalysis.LoadModeTypesInfo)
-}
