--- pkg/golinters/gosimple/gosimple_integration_test.go
@@ -1,19 +1,0 @@
-package gosimple
-
-import (
-	"testing"
-
-	"github.com/golangci/golangci-lint/test/testshared/integration"
-)
-
-func TestFromTestdata(t *testing.T) {
-	integration.RunTestdata(t)
-}
-
-func TestFix(t *testing.T) {
-	integration.RunFix(t)
-}
-
-func TestFixPathPrefix(t *testing.T) {
-	integration.RunFixPathPrefix(t)
-}
