+++ pkg/golinters/staticcheck/testdata/fix/in/gosimple.go
@@ -1,4 +1,5 @@
-//golangcitest:args -Egosimple
+//golangcitest:args -Estaticcheck
+//golangcitest:config_path testdata/gosimple.yml
 //golangcitest:expected_exitcode 0
 package testdata
