+++ pkg/golinters/staticcheck/testdata/fix/in/staticcheck.go
@@ -1,4 +1,5 @@
 //golangcitest:args -Estaticcheck
+//golangcitest:config_path testdata/staticcheck.yml
 //golangcitest:expected_exitcode 0
 package testdata
