+++ test/testdata/default_exclude.go
@@ -1,12 +1,12 @@
-//golangcitest:args -Estylecheck,revive --internal-cmd-test
+//golangcitest:args -Estaticcheck,revive --internal-cmd-test
 //golangcitest:config_path testdata/configs/default_exclude.yml
 
 /*Package testdata ...*/
 package testdata
 
-// InvalidFuncComment, both revive and stylecheck will complain about this, // want `exported: comment on exported function ExportedFunc1 should be of the form "ExportedFunc1 ..."`
+// InvalidFuncComment, both revive and staticcheck (stylecheck) will complain about this, // want `exported: comment on exported function ExportedFunc1 should be of the form "ExportedFunc1 ..."`
 // if include EXC0011, only the warning from revive will be ignored.
-// And only the warning from stylecheck will start with "ST1020".
+// And only the warning from staticcheck (stylecheck) will start with "ST1020".
 func ExportedFunc1() {
 }
 
@@ -16,6 +16,6 @@ func ExportedFunc1() {
 func ExportedFunc2() {
 }
 
-//nolint:stylecheck
+//nolint:staticcheck
 func IgnoreAll() { // want "exported: exported function IgnoreAll should have comment or be unexported"
 }
